sleep-assistant/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
pydantic>=2.10.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.32.0
certifi>=2024.07.04
slowapi>=0.1.9
